Enables user activation/deactivation
Adds functionality to activate and deactivate users via the admin panel.

This change introduces a new controller action to update the `is_active` status of a user.
It also updates the user model to include the `is_active` field as a mass assignable attribute.
The user index page is updated to include dropdown menu with options to activate/deactivate users,
and the report show page now displays the user associated with the report.